requirements.txt
# Phase 1: Detection & Notifications
feedparser>=6.0.0
pytrends>=4.9.0
newsapi-python>=0.2.7
python-telegram-bot>=21.0
python-dotenv>=1.0.0
schedule>=1.2.0
aiohttp>=3.9.0

# Phase 2: Article Generation
google-genai>=1.0.0
trafilatura>=1.6.0
newspaper3k>=0.2.8
markdown>=3.5.0

# Phase 3: WordPress Publishing
requests>=2.31.0
Pillow>=10.0.0
